darcs-hash:20070920112531-a1b90-f56a2bb314fa654963394fd54c3df6de143ed95e.gz

+++ projects/SATIrE/src/analyzer/astaccess/satire/ast_support_funcs.h
@@ -1,20 +1,20 @@
 // Copyright 2005,2006,2007 Markus Schordan, Gergo Barany
-// $Id: ast_support_funcs.h,v 1.4 2008-05-19 12:37:09 gergo Exp $
+// $Id: ast_support_funcs.h,v 1.1 2007-09-20 09:25:31 adrian Exp $
 
 #ifndef H_AST_SUPPORT_FUNCS
 #define H_AST_SUPPORT_FUNCS
 
-#include <satire_rose.h>
+#include <rose.h>
 
-SgType *typenum_to_type(unsigned long n);
-unsigned long type_to_typenum(SgType *type);
-std::string typenum_to_str(unsigned long n);
-SgExpression *exprnum_to_expr(unsigned long n);
-unsigned long expr_to_exprnum(SgExpression *expr);
-std::string exprnum_to_str(unsigned long n);
+SgType *typenum_to_type(int n);
+int type_to_typenum(SgType *type);
+std::string typenum_to_str(int n);
+SgExpression *exprnum_to_expr(int n);
+int expr_to_exprnum(SgExpression *expr);
+std::string exprnum_to_str(int n);
 SgType *expr_type(SgExpression *expr);
-unsigned long exprnum_typenum(unsigned long n);
+int exprnum_typenum(int n);
 bool is_subtype_of(SgClassType *a, SgClassType *b);
-bool is_subtypenum_of(unsigned long a, unsigned long b);
+bool is_subtypenum_of(int a, int b);
 
 #endif
